feat(web): 일정 URL 라우팅 분리
