+++ web/src/app/schedules/backlog/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { BacklogPanel } from '@/features/schedule/components/backlog-panel';
+
+export default function BacklogPage() {
+  return <BacklogPanel />;
+}
