+++ web/src/app/schedules/categories/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { CategoriesPanel } from '@/features/schedule/components/categories-panel';
+
+export default function CategoriesPage() {
+  return <CategoriesPanel />;
+}
